backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
httpx>=0.27.0,<1.0.0
html2text>=2024.2.26,<2027.0.0
python-dotenv>=1.0.1,<2.0.0
PyYAML>=6.0.2,<7.0.0
tiktoken>=0.8.0,<1.0.0
langchain>=1.0.0,<2.0.0
langchain-openai>=1.0.0,<2.0.0
langchain-community>=0.3.0,<1.0.0
# langchain-deepseek>=1.0.0,<2.0.0 #（optional, 需要deepseek的api key）
# langfuse (optional, for tracing)
langfuse>=2.0.0
psycopg[binary]>=3.2.0,<4.0.0
pytest>=8.3.0,<9.0.0
pypdf>=5.4.0,<6.0.0
openpyxl>=3.1.5,<4.0.0

# memory_module_v2: BM25 keyword retrieval + Chinese tokenizer
rank-bm25>=0.2.2
jieba>=0.42.1

# Postgres checkpointer (optional; 需 CHECKPOINTER=postgres)
langgraph-checkpoint-postgres>=2.0.0

# Chroma for memory_module_v1 vector store（optional）
chromadb>=0.4.0
langchain-chroma>=0.1.0

#llama-index-core、llama-index-embeddings-openai：tools/search_knowledge_tool.py 里在用，可以替换为其他向量库（optional）
# llama-index-core>=0.12.0,<0.13.0
# llama-index-embeddings-openai>=0.3.0,<0.4.0
